requirements.txt
django
djangorestframework
djangorestframework-jsonapi
django-filter
djangorestframework-simplejwt